Refactor orangearg/argument/miner/solver.py as a separate sub-module.
