+++ setup.cfg
@@ -68,6 +68,7 @@ dev =
     nbsphinx_link
     black
     pylint
+    jupyter
 publishing = 
     twine
     wheel
